pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]
